chore: Moved definition/enum/http-status-code.enum.mts to core/server/definition/enum/http-status-code.enum.mts.
