+++ packages/architectura/src/core/server/definition/enum/http-status-code.enum.mts
