+++ packages/architectura/src/core/server/server.mts
@@ -2,7 +2,6 @@ import { Server as HTTPServer, type RequestListener } from "node:http";
 
 import { Server as HTTPSServer } from "node:https";
 
-import { HTTPStatusCodeEnum } from "../../definition/enum/http-status-code.enum.mjs";
 
 
 
@@ -18,6 +17,8 @@ import { EndpointRegistry } from "../endpoint/endpoint.registry.mjs";
 import { ExecutionContext } from "../execution-context/execution-context.mjs";
 
 import { ExecutionContextRegistry } from "../execution-context/execution-context.registry.mjs";
+
+import { HTTPStatusCodeEnum } from "./definition/enum/http-status-code.enum.mjs";
 
 import { PortsEnum } from "./definition/enum/ports.enum.mjs";
